Moved various CRUD methods from StatusDashboard to EnvironmentController. RepeatException camel case.
