+++ src/coltools/EnvironmentTest.java
@@ -1,0 +1,23 @@
+package coltools;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Created by elliot on 30/07/2018.
+ */
+class EnvironmentTest {
+    @Test
+    void getBaseUrl() {
+    }
+
+    @Test
+    void getUserName() {
+    }
+
+    @Test
+    void getPassword() {
+    }
+
+}
