+++ src/coltools/RepeatException.java
@@ -3,12 +3,12 @@ package coltools;
 /**
  * Created by Sam Leese on 16/07/2018.
  */
-public class repeatException extends Exception {
-    public repeatException(String message) {
+public class RepeatException extends Exception {
+    public RepeatException(String message) {
         super("Base URL already exists in system");
     }
 
-    public repeatException() {
+    public RepeatException() {
 
     }
 }
